supra-xrbc-coi/scripts/requirements.txt
python-dotenv
psutil
fabric
boto3
pandas
